refactor(Describe): rewrite in TS and hooks
