+++ src/containers/Tenant/Schema/SchemaTree/SchemaTree.tsx
@@ -4,7 +4,7 @@ import {useDispatch} from 'react-redux';
 import {NavigationTree} from 'ydb-ui-components';
 
 import {setCurrentSchemaPath, getSchema, preloadSchemas} from '../../../../store/reducers/schema';
-import {getDescribe} from '../../../../store/reducers/describe';
+// import {getDescribe} from '../../../../store/reducers/describe';
 import {getSchemaAcl} from '../../../../store/reducers/schemaAcl';
 import type {EPathType, TEvDescribeSchemeResult} from '../../../../types/api/schema';
 
@@ -63,8 +63,9 @@ export function SchemaTree(props: SchemaTreeProps) {
 
         // FIXME: these dispatches seem to be useless, as data is requested in every tab anyway
         // But it needs to be checked more carefully
+        // For example, somewhere it fixes not sufficient data update inside componentDidUpdate
         dispatch(getSchema({path: activePath}));
-        dispatch(getDescribe({path: activePath}));
+        // dispatch(getDescribe({path: activePath}));
         dispatch(getSchemaAcl({path: activePath}));
     };
